backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
strands-agents[anthropic,openai,ollama,gemini,mistral]>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.9.0
seaborn>=0.13.0
scikit-learn>=1.5.0
scipy>=1.14.0
